requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langchain-pinecone
langchain[docarray]
docarray
pydantic==1.10.8
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
ffmpeg
git+https://github.com/openai/whisper.git
google-api-core
protoc-gen-openapiv2
